issue 310 : move heap store into its own package
