+++ impl/src/main/java/org/ehcache/internal/store/heap/BaseOnHeapKey.java
@@ -13,17 +13,30 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.ehcache.internal.store;
 
-import org.ehcache.spi.cache.Store.ValueHolder;
+package org.ehcache.internal.store.heap;
 
-interface OnHeapValueHolder<V> extends ValueHolder<V> {
-  
-  void setAccessTimeMillis(long accessTime);
-  
-  void setExpireTimeMillis(long expireTime);
-  
-  boolean isExpired(long now);
+abstract class BaseOnHeapKey<K> implements OnHeapKey<K> {
 
-  long getExpireTimeMillis();
+  private final int hashCode;
+
+  BaseOnHeapKey(int hashCode) {
+    this.hashCode = hashCode;
+  }
+
+  @Override
+  public int hashCode() {
+    return hashCode;
+  }
+
+  @Override
+  public boolean equals(Object obj) {
+    if (obj instanceof OnHeapKey) {
+      OnHeapKey<?> other = (OnHeapKey<?>) obj;
+      return getActualKeyObject().equals(other.getActualKeyObject());
+    }
+
+    return false;
+  }
+
 }
