+++ impl/src/main/java/org/ehcache/internal/store/heap/ByValueOnHeapValueHolder.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.ehcache.internal.store;
+package org.ehcache.internal.store.heap;
 
 import org.ehcache.exceptions.SerializerException;
 import org.ehcache.spi.cache.Store.ValueHolder;
