+++ impl/src/test/java/org/ehcache/internal/store/disk/DiskStoreSPITest.java
@@ -25,7 +25,7 @@ import org.ehcache.expiry.Expiry;
 import org.ehcache.internal.SystemTimeSource;
 import org.ehcache.internal.TimeSource;
 import org.ehcache.internal.serialization.JavaSerializationProvider;
-import org.ehcache.internal.store.OnHeapStore;
+import org.ehcache.internal.store.heap.OnHeapStore;
 import org.ehcache.internal.store.StoreFactory;
 import org.ehcache.internal.store.StoreSPITest;
 import org.ehcache.internal.store.disk.DiskStorageFactory.Element;
