+++ impl/src/test/java/org/ehcache/internal/store/heap/ByValueOnHeapKeyTest.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.ehcache.internal.store;
+package org.ehcache.internal.store.heap;
 
 import org.ehcache.internal.serialization.JavaSerializer;
 import org.junit.Test;
