+++ impl/src/test/java/org/ehcache/internal/store/heap/OnHeapStoreByValueSPITest.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.ehcache.internal.store;
+package org.ehcache.internal.store.heap;
 
 import org.ehcache.config.EvictionPrioritizer;
 import org.ehcache.config.EvictionVeto;
@@ -24,6 +24,10 @@ import org.ehcache.expiry.Expiry;
 import org.ehcache.internal.HeapResourceCacheConfiguration;
 import org.ehcache.internal.SystemTimeSource;
 import org.ehcache.internal.TimeSource;
+import org.ehcache.internal.serialization.JavaSerializer;
+import org.ehcache.internal.store.StoreFactory;
+import org.ehcache.internal.store.StoreSPITest;
+import org.ehcache.internal.store.heap.service.OnHeapStoreServiceConfig;
 import org.ehcache.spi.ServiceLocator;
 import org.ehcache.spi.ServiceProvider;
 import org.ehcache.spi.cache.Store;
@@ -31,13 +35,13 @@ import org.ehcache.spi.service.ServiceConfiguration;
 import org.junit.Before;
 
 /**
- * Test the {@link org.ehcache.internal.store.OnHeapStore} compliance to the
+ * Test the {@link org.ehcache.internal.store.heap.OnHeapStore} compliance to the
  * {@link org.ehcache.spi.cache.Store} contract.
  *
  * @author Aurelien Broszniowski
  */
 
-public class OnHeapStoreByRefSPITest extends StoreSPITest<String, String> {
+public class OnHeapStoreByValueSPITest extends StoreSPITest<String, String> {
 
   private StoreFactory<String, String> storeFactory;
 
@@ -48,30 +52,35 @@ public class OnHeapStoreByRefSPITest extends StoreSPITest<String, String> {
 
   @Before
   public void setUp() {
+    
     storeFactory = new StoreFactory<String, String>() {
 
       @Override
       public Store<String, String> newStore(final Store.Configuration<String, String> config) {
-        return new OnHeapStore<String, String>(config, SystemTimeSource.INSTANCE, false, null, null);
+        return new OnHeapStore<String, String>(config, SystemTimeSource.INSTANCE, true, new JavaSerializer<String>(getClass().getClassLoader()), new JavaSerializer<String>(getClass().getClassLoader()));
       }
 
       @Override
       public Store<String, String> newStore(final Store.Configuration<String, String> config, TimeSource timeSource) {
-        return new OnHeapStore<String, String>(config, timeSource, false, null, null);
+        return new OnHeapStore<String, String>(config, timeSource, true, new JavaSerializer<String>(getClass().getClassLoader()), new JavaSerializer<String>(getClass().getClassLoader()));
       }
 
       @Override
       public Store.ValueHolder<String> newValueHolder(final String value) {
-        return new ByRefOnHeapValueHolder<String>(value, SystemTimeSource.INSTANCE.getTimeMillis());
+        return new ByValueOnHeapValueHolder<String>(value, SystemTimeSource.INSTANCE.getTimeMillis(), new JavaSerializer<String>(getClass().getClassLoader()));
       }
 
       @Override
       public Store.Provider newProvider() {
-        return new OnHeapStore.Provider();
+        Store.Provider service = new OnHeapStore.Provider();
+        service.start(null, new ServiceLocator());
+        return service;
       }
 
       @Override
-      public Store.Configuration<String, String> newConfiguration(final Class<String> keyType, final Class<String> valueType, final Comparable<Long> capacityConstraint, final EvictionVeto<? super String, ? super String> evictionVeto, final EvictionPrioritizer<? super String, ? super String> evictionPrioritizer) {
+      public Store.Configuration<String, String> newConfiguration(
+          final Class<String> keyType, final Class<String> valueType, final Comparable<Long> capacityConstraint,
+          final EvictionVeto<? super String, ? super String> evictionVeto, final EvictionPrioritizer<? super String, ? super String> evictionPrioritizer) {
         return new StoreConfigurationImpl<String, String>(keyType, valueType, capacityConstraint,
             evictionVeto, evictionPrioritizer, ClassLoader.getSystemClassLoader(), Expirations.noExpiration());
       }
@@ -97,7 +106,7 @@ public class OnHeapStoreByRefSPITest extends StoreSPITest<String, String> {
 
       @Override
       public ServiceConfiguration<?>[] getServiceConfigurations() {
-        return new ServiceConfiguration[] { new HeapResourceCacheConfiguration(100) };
+        return new ServiceConfiguration[] { new HeapResourceCacheConfiguration(100), new OnHeapStoreServiceConfig().storeByValue(true)};
       }
 
       @Override
